Ruler: Unparent Popover in dtor so !use-after-free
There was a warning printed about trying to finalise a widget that still
had children, but its :name printed in the warning was gibberish. ASAN
showed that it was the Ruler::_popover. This might mean that popovers
that are parented by a widget 'manually' must be unparented in the dtor
and/or in a signal_destroy() handler. The former seems to suffice here,
but I do both as documented for signal_destroy(), in case that's needed.

Other popover-having widgets probably need similar treatment, but let's
approach them on a case-by-case basis; I'm onto a DIFFERENT crash now...