Replace generic unparent(), w/ specific remove()s
gtk_widget_unparent() is for widget implementations only, not a way to
remove arbitrary widgets from whatever parent they happen to be in, as
gtk_container_remove() used to be. Just unparent()ing does not suffice
for more complicated 'container' widgets in GTK4, which do other stuff
in their own, unrelated, non-virtual remove() methods. So we shouldn't
be using unparent() except where we ourselves previously set_parent().

This commit Makes It So.