Add Inkscape::ustring::format() that uses the "classic" ("C") locale.
Gtk4 ensures that the C++ locale matches the C locale which was
already set to the user's preferred local. Glib::ustring::format()
uses the C++ locale in formatting. This causes trouble when numeric
output is expected to match that of the classic "C" locale (e.g. when
formatting an 8 digit hex number). This commit adds a function, based
on the Glib::ustring::format() code which ensures that the "classic"
locale is used in formatting (without changing the global locale). It
uses this new function where necessary to ensure formatted text is as
expected.

Inkscape::ustring::format_classic() can be replaced by std::format()
once the latter is supported by Apple Clang.